refactor: Extract ViewModel extension function
This commit extracts the `launchCatching` extension function from `MyListViewModel.kt` into its own file, `ViewModelExtensions.kt`, under the `com.branwen.mal.utils` package.

The `launchCatching` function provides a convenient way to launch a coroutine within a `ViewModel`'s `viewModelScope` on the `Dispatchers.IO` context, with built-in error handling using `runCatching` and logging via `Timber`. It also supports an optional `post` block that executes after the main block, with its own error handling.

This change improves code organization and reusability by making the `launchCatching` function accessible to other ViewModels.
